pyproject.toml
[project]
name = "robot-stack"
version = "0.1.0"
description = "Tools to inspect and build Opentrons robots"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "ruff>=0.11.5",
    "rust-just>=1.40.0",
    "semver>=3.0.4",
    "types-pyyaml>=6.0.12.20250402",
]

[tool.black]
line-length = 130
exclude = '''
(
  /(
      \.git
    | \.venv
    | \.github
    | buildroot
    | opentrons
    | ot3-firmware
    | oe-core
  )/
)
'''

[tool.ruff]
line-length = 130
exclude = [
    ".git",
    ".venv",
    ".github",
    "buildroot",
    "opentrons",
    "ot3-firmware",
    "oe-core",
]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
# Exclude top‑level dirs and common non‑code folders
exclude = '''
(?x)                      # allow comments & whitespace
^
(                          # any of these names at the repo root
    buildroot
  | opentrons
  | ot3-firmware
  | oe-core
  | \.git
  | \.venv
  | \.github
)
(/|$)                      # followed by a slash or end‑of‑string
'''
